refactor: organize utility functions and improve environment variable handling
