+++ src/cleanup/mod.rs
@@ -117,18 +117,18 @@ fn prune_project_unused() {
         let nodejs_dist_path = i.location.to_string() + "/dist";
         lrncore::path::change_work_dir(&i.location);
         if lrncore::path::path_exists(&node_module_path) {
-            utils::sys::rm_command(node_module_path);
-            utils::sys::rm_command(nodejs_dist_path);
+            utils::fsys::rm_command(node_module_path);
+            utils::fsys::rm_command(nodejs_dist_path);
         }
         // Golang
         let bin_folder = i.location.to_string() + "/bin/";
         if lrncore::path::path_exists(&bin_folder) {
-            utils::sys::rm_command(bin_folder);
+            utils::fsys::rm_command(bin_folder);
         }
         // Rust
         let target_folder = i.location.to_string() + "/target";
         if lrncore::path::path_exists(&target_folder) {
-            utils::sys::rm_command(target_folder);
+            utils::fsys::rm_command(target_folder);
         }
     }
 }
