+++ src/utils/env.rs
@@ -1,0 +1,9 @@
+use std::env;
+
+pub fn get_nyx_env_var() -> String {
+    let env_var = "NYX";
+    match env::var(env_var) {
+        Ok(v) => return v,
+        Err(e) => panic!("${} is not set ({})", env_var, e),
+    }
+}
